requirements.txt
python-telegram-bot==21.10
playwright==1.50.0
requests==2.32.3
python-dotenv==1.0.1
